worker/translib-requirements.txt
nltk
flask-cors
nltk_tgrep
unidecode